Creates ToDo.DB, changes read_file and write_file to all and save
